requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.0
python-dotenv==1.0.0
pydantic==2.5.0
googlesearch-python==1.2.3
feedparser==6.0.10
pandas==2.1.3
markdown==3.5.1